% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Blog.locale}
\alias{Blog.locale}
\title{Blog.locale Object}
\usage{
Blog.locale(country = NULL, language = NULL, variant = NULL)
}
\arguments{
\item{country}{The country this blog's locale is set to}

\item{language}{The language this blog is authored in}

\item{variant}{The language variant this blog is authored in}
}
\value{
Blog.locale object
}
\description{
Blog.locale Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The locale this Blog is set to.
}
\seealso{
Other Blog functions: \code{\link{Blog.pages}},
  \code{\link{Blog.posts}}, \code{\link{Blog}}
}

