% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Blog.pages}
\alias{Blog.pages}
\title{Blog.pages Object}
\usage{
Blog.pages(selfLink = NULL, totalItems = NULL)
}
\arguments{
\item{selfLink}{The URL of the container for pages in this blog}

\item{totalItems}{The count of pages in this blog}
}
\value{
Blog.pages object
}
\description{
Blog.pages Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The container of pages in this blog.
}
\seealso{
Other Blog functions: \code{\link{Blog.locale}},
  \code{\link{Blog.posts}}, \code{\link{Blog}}
}

