% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\docType{package}
\name{blogger_googleAuthR}
\alias{blogger_googleAuthR}
\alias{blogger_googleAuthR-package}
\title{Blogger API
API for access to the data within Blogger.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:29:17
filename: /Users/mark/dev/R/autoGoogleAPI/googlebloggerv2.auto/R/blogger_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}
}
