% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{BlogList}
\alias{BlogList}
\title{BlogList Object}
\usage{
BlogList(items = NULL)
}
\arguments{
\item{items}{The list of Blogs this user has Authorship or Admin rights over}
}
\value{
BlogList object
}
\description{
BlogList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

