% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Comment.author.image}
\alias{Comment.author.image}
\title{Comment.author.image Object}
\usage{
Comment.author.image(url = NULL)
}
\arguments{
\item{url}{The comment creator's avatar URL}
}
\value{
Comment.author.image object
}
\description{
Comment.author.image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The comment creator's avatar.
}
\seealso{
Other Comment functions: \code{\link{Comment.author}},
  \code{\link{Comment.blog}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.post}}, \code{\link{Comment}}
}
