% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Comment.author}
\alias{Comment.author}
\title{Comment.author Object}
\usage{
Comment.author(Comment.author.image = NULL, displayName = NULL, id = NULL,
  image = NULL, url = NULL)
}
\arguments{
\item{Comment.author.image}{The \link{Comment.author.image} object or list of objects}

\item{displayName}{The display name}

\item{id}{The identifier of the Comment creator}

\item{image}{The comment creator's avatar}

\item{url}{The URL of the Comment creator's Profile page}
}
\value{
Comment.author object
}
\description{
Comment.author Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The author of this Comment.
}
\seealso{
Other Comment functions: \code{\link{Comment.author.image}},
  \code{\link{Comment.blog}},
  \code{\link{Comment.inReplyTo}},
  \code{\link{Comment.post}}, \code{\link{Comment}}
}

