% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Comment.post}
\alias{Comment.post}
\title{Comment.post Object}
\usage{
Comment.post(id = NULL)
}
\arguments{
\item{id}{The identifier of the post containing this comment}
}
\value{
Comment.post object
}
\description{
Comment.post Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about the post containing this comment.
}
\seealso{
Other Comment functions: \code{\link{Comment.author.image}},
  \code{\link{Comment.author}}, \code{\link{Comment.blog}},
  \code{\link{Comment.inReplyTo}}, \code{\link{Comment}}
}

