% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{CommentList}
\alias{CommentList}
\title{CommentList Object}
\usage{
CommentList(items = NULL, nextPageToken = NULL, prevPageToken = NULL)
}
\arguments{
\item{items}{The List of Comments for a Post}

\item{nextPageToken}{Pagination token to fetch the next page, if one exists}

\item{prevPageToken}{Pagination token to fetch the previous page, if one exists}
}
\value{
CommentList object
}
\description{
CommentList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
