% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{comments.list}
\alias{comments.list}
\title{Retrieves the comments for a blog, possibly filtered.}
\usage{
comments.list(blogId, postId, fetchBodies = NULL, maxResults = NULL,
  pageToken = NULL, startDate = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch comments from}

\item{postId}{ID of the post to fetch posts from}

\item{fetchBodies}{Whether the body content of the comments is included}

\item{maxResults}{Maximum number of comments to include in the result}

\item{pageToken}{Continuation token if request is paged}

\item{startDate}{Earliest date of comment to fetch, a date-time with RFC 3339 formatting}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/2.0/json/getting_started}{Google Documentation}
}
