% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Page.blog}
\alias{Page.blog}
\title{Page.blog Object}
\usage{
Page.blog(id = NULL)
}
\arguments{
\item{id}{The identifier of the blog containing this page}
}
\value{
Page.blog object
}
\description{
Page.blog Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about the blog containing this Page.
}
\seealso{
Other Page functions: \code{\link{Page.author.image}},
  \code{\link{Page.author}}, \code{\link{Page}}
}
