% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{PageList}
\alias{PageList}
\title{PageList Object}
\usage{
PageList(items = NULL)
}
\arguments{
\item{items}{The list of Pages for a Blog}
}
\value{
PageList object
}
\description{
PageList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

