% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Post.author}
\alias{Post.author}
\title{Post.author Object}
\usage{
Post.author(Post.author.image = NULL, displayName = NULL, id = NULL,
  image = NULL, url = NULL)
}
\arguments{
\item{Post.author.image}{The \link{Post.author.image} object or list of objects}

\item{displayName}{The display name}

\item{id}{The identifier of the Post creator}

\item{image}{The Post author's avatar}

\item{url}{The URL of the Post creator's Profile page}
}
\value{
Post.author object
}
\description{
Post.author Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The author of this Post.
}
\seealso{
Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.blog}}, \code{\link{Post.replies}},
  \code{\link{Post}}
}
