% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Post}
\alias{Post}
\title{Post Object}
\usage{
Post(Post.author = NULL, Post.author.image = NULL, Post.blog = NULL,
  Post.replies = NULL, author = NULL, blog = NULL, content = NULL,
  id = NULL, labels = NULL, published = NULL, replies = NULL,
  selfLink = NULL, title = NULL, updated = NULL, url = NULL)
}
\arguments{
\item{Post.author}{The \link{Post.author} object or list of objects}

\item{Post.author.image}{The \link{Post.author.image} object or list of objects}

\item{Post.blog}{The \link{Post.blog} object or list of objects}

\item{Post.replies}{The \link{Post.replies} object or list of objects}

\item{author}{The author of this Post}

\item{blog}{Data about the blog containing this Post}

\item{content}{The content of the Post}

\item{id}{The identifier of this Post}

\item{labels}{The list of labels this Post was tagged with}

\item{published}{RFC 3339 date-time when this Post was published}

\item{replies}{The container of comments on this Post}

\item{selfLink}{The API REST URL to fetch this resource from}

\item{title}{The title of the Post}

\item{updated}{RFC 3339 date-time when this Post was last updated}

\item{url}{The URL where this Post is displayed}
}
\value{
Post object
}
\description{
Post Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.author}}, \code{\link{Post.blog}},
  \code{\link{Post.replies}}
}

