% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{Post.replies}
\alias{Post.replies}
\title{Post.replies Object}
\usage{
Post.replies(selfLink = NULL, totalItems = NULL)
}
\arguments{
\item{selfLink}{The URL of the comments on this post}

\item{totalItems}{The count of comments on this post}
}
\value{
Post.replies object
}
\description{
Post.replies Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The container of comments on this Post.
}
\seealso{
Other Post functions: \code{\link{Post.author.image}},
  \code{\link{Post.author}}, \code{\link{Post.blog}},
  \code{\link{Post}}
}
