% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_functions.R
\name{posts.list}
\alias{posts.list}
\title{Retrieves a list of posts, possibly filtered.}
\usage{
posts.list(blogId, fetchBodies = NULL, maxResults = NULL,
  pageToken = NULL, startDate = NULL)
}
\arguments{
\item{blogId}{ID of the blog to fetch posts from}

\item{fetchBodies}{Whether the body content of posts is included}

\item{maxResults}{Maximum number of posts to fetch}

\item{pageToken}{Continuation token if the request is paged}

\item{startDate}{Earliest post date to fetch, a date-time with RFC 3339 formatting}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/blogger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/blogger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/blogger/docs/2.0/json/getting_started}{Google Documentation}
}

