% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{User.blogs}
\alias{User.blogs}
\title{User.blogs Object}
\usage{
User.blogs(selfLink = NULL)
}
\arguments{
\item{selfLink}{The URL of the Blogs for this user}
}
\value{
User.blogs object
}
\description{
User.blogs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The container of blogs for this user.
}
\seealso{
Other User functions: \code{\link{User.locale}},
  \code{\link{User}}
}

