% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogger_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(User.blogs = NULL, User.locale = NULL, about = NULL, blogs = NULL,
  created = NULL, displayName = NULL, id = NULL, locale = NULL,
  selfLink = NULL, url = NULL)
}
\arguments{
\item{User.blogs}{The \link{User.blogs} object or list of objects}

\item{User.locale}{The \link{User.locale} object or list of objects}

\item{about}{Profile summary information}

\item{blogs}{The container of blogs for this user}

\item{created}{The timestamp of when this profile was created, in seconds since epoch}

\item{displayName}{The display name}

\item{id}{The identifier for this User}

\item{locale}{This user's locale}

\item{selfLink}{The API REST URL to fetch this resource from}

\item{url}{The user's profile page}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other User functions: \code{\link{User.blogs}},
  \code{\link{User.locale}}
}
